requirements.txt
absl-py==0.7.1
appnope==0.1.0
argh==0.26.2
astor==0.8.0
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
baselines==0.1.5
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
-e git+https://github.com/openai/coinrun.git@523704f3a203dcaad84caf96ea92799452dc902f#egg=coinrun
colorama==0.4.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.0
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.11
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
geomloss==0.2.1
gitdb2==2.0.5
GitPython==2.1.11
google-pasta==0.1.7
gql==0.1.0
graphql-core==2.2
grpcio==1.23.0
gym==0.14.0
gym-minigrid==0.0.4
h5py==2.9.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.19
ipykernel==5.1.2
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.2.0
ml-metrics==0.1.4
more-itertools==7.2.0
mpi4py==3.0.2
natsort==6.0.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numpy==1.16.4
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.1.0.25
packaging==19.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.0
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
POT==0.5.1
progressbar2==3.43.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyarrow==0.14.1
pycparser==2.19
pyglet==1.3.2
Pygments==2.4.2
pyparsing==2.4.2
PyQt5==5.12.3
PyQt5-sip==4.19.17
pyrsistent==0.15.4
pytest==5.0.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.1.0
qtconsole==4.5.4
ray==0.7.2
redis==3.3.1
requests==2.22.0
Rx==1.6.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
sentry-sdk==0.10.0
shortuuid==0.5.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
stable-baselines==2.7.0
subprocess32==3.5.4
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wandb
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.5.2
